upd-for-pre02

+++ .py/psh.py
@@ -1,17 +1,22 @@
 # -*- coding: utf-8 -*-
 
 import urllib,sys,os
-
-def getf(l):
-    print 'fetch %s' % l
-    return urllib.urlopen(l).read()
-
 cfg = open('../config.cfg').read().splitlines()
 
-website = '%sz/push' % cfg[1]
-auth = cfg[0]
+def postf(s):
+    data = urllib.urlencode({'tmsg': s,'pauth': cfg[0]})
+    #{"Content-type": "application/x-www-form-urlencoded"}
+    u = urllib.urlopen(cfg[1] + 'z/point', data)
+    return u.read()
 
 for ml in sys.argv[1:]:
     txt = open(ml).read()
-    out = getf('%s/%s/%s' % (website, auth, txt))
-    os.remove(ml)
+    out = postf(txt)
+    if out.startswith('msg ok'):
+        os.remove(ml)
+    elif out == 'msg big!':
+        print 'ERROR: very big message (limit: 64k)!'
+    elif out == 'auth error!':
+        print 'ERROR: AUTH!'
+    else:
+        print 'ERROR: WHY?'
